# app/public/openapi.yaml
openapi: 3.0.0
info:
  title: 'Gestor de Condomínios'
  version: '0.1'
paths:
...
  /api/veiculo:
    post:
      tags:
        - Veículo
      summary: 'Realiza o cadastro completo de veículo'
      operationId: 7ba9160c41dad34d8c88902b75ac8e9c
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                categoria:
                  type: integer
                cor:
                  type: integer
                marca:
                  type: integer
                proprietario:
                  type: integer
                propriedade:
                  type: integer
                modelo:
                  type: string
                placa:
                  type: string
                ano:
                  type: string
              type: object
      responses:
        '200':
          description: 'Requisição bem-sucedida'
        '400':
          description: 'Requisição inválida, dados incorretos ou faltando parâmetros'
        '500':
          description: 'Erro interno do servidor'
      security:
        -
          api_key: []
  /api/veiculo/listagem:
    get:
      tags:
        - Veículo
      summary: 'Listagem de Veículo do Sistema'
      operationId: f5fb81bb5be407f1f64275af8fbe385c
      parameters:
        -
          name: nome
          in: query
          description: Nome
          required: false
          schema:
            type: string
        -
          name: placa
          in: query
          description: Placa
          required: false
          schema:
            type: string
        -
          name: modelo
          in: query
          description: Modelo
          required: false
          schema:
            type: integer
        -
          name: categoria
          in: query
          description: Categoria
          required: false
          schema:
            type: string
        -
          name: pagina
          in: query
          description: Página
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'Requisição bem-sucedida'
        '400':
          description: 'Requisição inválida, dados incorretos ou faltando parâmetros'
        '500':
          description: 'Erro interno do servidor'
      security:
        -
          api_key: []
  '/api/veiculo/{id}':
    get:
      tags:
        - Veículo
      summary: 'Consultar Veículo'
      operationId: 2a988ada611c4a04f1eec8ef111dde46
      parameters:
        -
          name: id
          in: path
          description: id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'Requisição bem-sucedida'
        '400':
          description: 'Requisição inválida, dados incorretos ou faltando parâmetros'
        '500':
          description: 'Erro interno do servidor'
      security:
        -
          api_key: []
    put:
      tags:
        - Veículo
      summary: 'Atualiza o cadastro completo de veículo'
      operationId: 507bbd7b9560785278db7402ba89a4d5
      parameters:
        -
          name: id
          in: path
          description: Id
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                categoria:
                  type: integer
                cor:
                  type: integer
                marca:
                  type: integer
                proprietario:
                  type: integer
                propriedade:
                  type: integer
                modelo:
                  type: string
                placa:
                  type: string
                ano:
                  type: string
              type: object
      responses:
        '200':
          description: 'Requisição bem-sucedida'
        '400':
          description: 'Requisição inválida, dados incorretos ou faltando parâmetros'
        '500':
          description: 'Erro interno do servidor'
      security:
        -
          api_key: []
    delete:
      tags:
        - Veículo
      summary: 'Excluir Veículo do Sistema'
      operationId: 3c84e4e6f3f21d85999f3dad735fda73
      parameters:
        -
          name: id
          in: path
          description: Id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'Requisição bem-sucedida'
        '400':
          description: 'Requisição inválida, dados incorretos ou faltando parâmetros'
        '500':
          description: 'Erro interno do servidor'
      security:
        -
          api_key: []
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      name: Authorization
      in: header
...